Use I²C repeated start condition in WriteRead trait
The embedded_hal WriteRead trait contract requires that a repeated
start is used between the write and read parts of the transaction.
This is not the same thing as sending a write transaction followed
by a read transaction and a "stop start" sequence between them.

For example, some devices such as the NXP MMA8652FC 3-axis
accelerometer will not see successive write and read transactions
as a combined one unless the repeated start is used.