puf: Support started PUF peripherals
The PUF might already be started when we try to access it from the
firmware, e. g. because the ROM started it.  This patch adds a
Puf::try_start method that starts the PUF if allowed or otherwise just
returns a started Puf instance, assuming that it has been started
already.

Fixes: https://github.com/lpc55/lpc55-hal/issues/64